connect & spawn actionPawn

+++ ActionServer/ActionServer/Vector3.cpp
@@ -1,21 +1,21 @@
-#include "Vector3.hpp"
-#include "Quaternion.hpp"
+#include "Vector3.h"
+#include "Quaternion.h"
 
 
 Quaternion Vector3::ToQuaternion() const
 {
 	const float DEG_TO_RAD = PI / ( 180.f );
-	double x = X*DEG_TO_RAD;
-	double y = Y*DEG_TO_RAD;
-	double z = Z*DEG_TO_RAD;
+	float x = X*DEG_TO_RAD;
+	float y = Y*DEG_TO_RAD;
+	float z = Z*DEG_TO_RAD;
 
 
-	double cx = cos( x * 0.5 );
-	double cy = cos( y * 0.5 );
-	double cz = cos( z * 0.5 );
-	double sx = sin( x * 0.5 );
-	double sy = sin( y * 0.5 );
-	double sz = sin( z * 0.5 );
+	float cx = (float)cos( x * 0.5 );
+	float cy = (float)cos( y * 0.5 );
+	float cz = (float)cos( z * 0.5 );
+	float sx = (float)sin( x * 0.5 );
+	float sy = (float)sin( y * 0.5 );
+	float sz = (float)sin( z * 0.5 );
 	Quaternion q;
 
 	//q.X = cx * sy * sz + cy * cz * sx;
